requirements.txt
flask
flask-compress
supabase
python-dotenv
google-genai
